исправления по домашней работе \Stream API и Optional\
